--- src/main/java/com/sangchu/batch/patch/entity/BaseTimeEntity.java
@@ -1,18 +1,0 @@
-package com.sangchu.batch.patch.entity;
-
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-@Getter
-public abstract class BaseTimeEntity {
-    
-    @CreatedDate
-    private LocalDateTime createdAt;
-}
