+++ src/main/java/com/sangchu/elasticsearch/repository/RecentIndexingDocRepository.java
@@ -1,0 +1,8 @@
+package com.sangchu.elasticsearch.repository;
+
+import com.sangchu.elasticsearch.entity.RecentIndexingDoc;
+import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
+
+public interface RecentIndexingDocRepository extends ElasticsearchRepository<RecentIndexingDoc, String> {
+
+}
