+++ src/main/java/com/sangchu/elasticsearch/repository/StoreSearchDocRepository.java
@@ -1,0 +1,11 @@
+package com.sangchu.elasticsearch.repository;
+
+import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
+import org.springframework.stereotype.Repository;
+
+import com.sangchu.elasticsearch.entity.StoreSearchDoc;
+
+@Repository
+public interface StoreSearchDocRepository extends ElasticsearchRepository<StoreSearchDoc, String>{
+
+}
