+++ src/main/java/com/sangchu/global/exception/custom/CustomException.java
@@ -2,12 +2,20 @@ package com.sangchu.global.exception.custom;
 
 import com.sangchu.global.util.statuscode.ApiStatus;
 
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
-@AllArgsConstructor
 public class CustomException extends RuntimeException {
 	private final ApiStatus status;
+
+	public CustomException(ApiStatus status) {
+		super(status.getMessage());
+		this.status = status;
+	}
+
+	public CustomException(ApiStatus status, String message) {
+		super(message);
+		this.status = status;
+	}
 
 }
